// src/lib/constants/notifications.enum.ts
export enum CategoryEnums {
    you_received_a_connection_invite = "you_received_a_connection_invite",
    you_sent_a_connection_invite = "you_sent_a_connection_invite",
    your_connection_invite_was_accepted = "your_connection_invite_was_accepted",
    you_received_a_message_from_marketplace = "you_received_a_message_from_marketplace",
    you_received_a_message_from_investment = "you_received_a_message_from_investment",
    you_received_a_message_from_outreach_hub = "you_received_a_message_from_outreach_hub",
    you_received_a_message_from_chat = "you_received_a_message_from_chat",
    happy_birthday = "happy_birthday",
}

export enum PostEnums {
    your_blog_post_published = "your_blog_post_published",
    your_investment_post_published = "your_investment_post_published",
    your_marketplace_post_published = "your_marketplace_post_published",
    your_outreach_hub_post_published = "your_outreach_hub_post_published",
}